Creation of login_helper.py with verify_token function
